Remove legacy fuzzes with go native fuzz, remove OSS-Fuzz build script until package is on OSS-Fuzz
