--- .github/workflows/fuzz_build.yml
@@ -1,31 +1,0 @@
-name: Build Fuzz
-
-on:
-  pull_request:
-  push:
-    branches:
-      - main
-
-jobs:
-  build_fuzz:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        go: ["1.19"]
-      fail-fast: false
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-go@v3
-        with:
-          go-version: ${{ matrix.go }}
-      - uses: technote-space/get-diff-action@v4
-        with:
-          PATTERNS: |
-            **/**.go
-            go.mod
-            go.sum
-      - name: Install dependencies
-        run: go get -u github.com/dvyukov/go-fuzz/go-fuzz github.com/dvyukov/go-fuzz/go-fuzz-build
-      - name: build fuzz
-        run: go-fuzz-build
-        working-directory: fuzz
