--- oss-fuzz-build.sh
@@ -1,6 +1,0 @@
-#!/bin/bash -eu
-
-export FUZZ_ROOT="github.com/pokt-network/smt"
-
-compile_go_fuzzer "$FUZZ_ROOT"/fuzz Fuzz fuzz_basic_op fuzz
-compile_go_fuzzer "$FUZZ_ROOT"/fuzz/delete Fuzz fuzz_delete fuzz
